+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Release instructions to README
+
 ## [0.20.4] - 2025-06-07
 
 ### Fixed
@@ -44,6 +47,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Bump dependencies
+
+### Fixed
+- Fix type issues
 
 ## [0.19.2] - 2025-03-23